% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{elist_randomize}
\alias{elist_randomize}
\title{Generate a random normally distributed version of an EList object.
Weights are used to choose the standard deviation of values generated.}
\usage{
elist_randomize(elist)
}
\description{
Generate a random normally distributed version of an EList object.
Weights are used to choose the standard deviation of values generated.
}
