% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_shift.R
\name{end_shift_pipeline}
\alias{end_shift_pipeline}
\title{End shift statistics from tail-tools pipeline output.}
\usage{
end_shift_pipeline(path, condition, group = NULL, select = NULL,
  ci = 0.95, fdr = T, edger = T, limma = T, antisense = T,
  colliders = T, non_utr = T, min_reads = 10,
  title = "End-shift test", fdr_max_permute = 1000)
}
\arguments{
\item{fdr}{Produce permutation based q values, can be very slow.

Samples with NA in condition will be omitted.
Samples can also be selected using select=, which should contain a logical vector of samples to retain.}
}
\description{
End shift statistics from tail-tools pipeline output.
}
