% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_shift_rnaseq.R
\name{end_shift_rnaseq}
\alias{end_shift_rnaseq}
\title{End-shift test for RNA-Seq data}
\usage{
end_shift_rnaseq(samples, utrs, extended_utrs = NULL, exons = NULL,
  ci = 0.95, min_min_reads = 20)
}
\arguments{
\item{samples}{should be a data frame with columns: name = name of sample, condition = logical vector with NA for samples to ignore, bigwig filenames giving depth of coverage cover_fwd, cover_rev.}

\item{utrs}{a GRanges of 3' UTRS. Should have metadata columns: ID, Name, gene_id, gene, description.}

\item{extended_utrs}{if given tries extending 3' UTRS beyond the annotated end point. A GRanges of validly extended 3' UTRS, in the same order as utrs. Suggestion is UTRs extended up to 10kb but not into the following gene on the same strand.}

\item{exons}{if given, restricts testing to exonic regions. A GRanges of transcript exons, which should have metadata column Parent corresponding to the ID in the utrs parameter.}
}
\description{
End-shift test for RNA-Seq data
}
