% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_browser.R
\name{load_genome_features}
\alias{load_genome_features}
\title{Load features needed by plot_genome().}
\usage{
load_genome_features(reference_dir)
}
\description{
Load features needed by plot_genome().
}

