% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pipeline_samples}
\alias{pipeline_samples}
\title{Get sample information from a Tail Tools pipeline output directory}
\usage{
pipeline_samples(pipeline_dir)
}
\description{
Get sample information from a Tail Tools pipeline output directory
}

