% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_browser.R
\name{plot_genome}
\alias{plot_genome}
\title{Produce a depth of coverage plot of a set of samples.}
\usage{
plot_genome(features, samples, pos, marks = c())
}
\description{
Produce a depth of coverage plot of a set of samples.
}
