% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_table.R
\name{prioritize_columns}
\alias{prioritize_columns}
\title{Put most important columns first in a data frame, for display}
\usage{
prioritize_columns(df, columns)
}
\description{
Put most important columns first in a data frame, for display
}

