% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{samples_group_tags}
\alias{samples_group_tags}
\title{Augment sample data frame with a column grouping several tags (logical columns) into a factor column}
\usage{
samples_group_tags(samples, group_name, tags, treatment = TRUE)
}
\description{
If treatment=FALSE, contrasts is set to "contr.sum".
}
