% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_end_shift.R
\name{shiny_end_shift}
\alias{shiny_end_shift}
\title{Show a report about an end shift analysis.}
\usage{
shiny_end_shift(result)
}
\arguments{
\item{result}{output from end_shift()}
}
\description{
Show a report about an end shift analysis.
}

