% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_end_shift.R
\name{shiny_end_shift_pipeline}
\alias{shiny_end_shift_pipeline}
\title{Perform end shift analyses on pipeline output and report results.}
\usage{
shiny_end_shift_pipeline(tests, cache_prefix = "cache_")
}
\arguments{
\item{tests}{Named list of lists giving parameters to end_shift_pipeline. Each name should be unique.}
}
\description{
Perform end shift analyses on pipeline output and report results.
}

