% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_end_shift_rnaseq.R
\name{shiny_end_shift_rnaseq_multiple}
\alias{shiny_end_shift_rnaseq_multiple}
\title{Interface to multiple RNA-Seq end-shift tests.}
\usage{
shiny_end_shift_rnaseq_multiple(samples, tests, references,
  cache_prefix = "cache", prefix = "",
  title = "RNA-Seq end-shift test")
}
\arguments{
\item{samples}{A data frame of samples. Should have columns: name = name of sample, bigwig files: cover_fwd, cover_rev, span_fwd, span_rev.}

\item{references}{A named character vector of reference directories created with "tail-tools make-rnaseq-reference:".}

\item{conditions}{A list of condition vectors.}
}
\description{
Interface to multiple RNA-Seq end-shift tests.
}
