% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_featureset.R
\name{shiny_featureset}
\alias{shiny_featureset}
\title{Display information about a subset of rows in a counts file}
\usage{
shiny_featureset(tc, fs = NULL, species = NULL, is_peaks = FALSE,
  prefix = "")
}
\description{
Display information about a subset of rows in a counts file
}

