% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mpat.R
\name{shiny_mpat}
\alias{shiny_mpat}
\title{Shiny report for mPAT results}
\usage{
shiny_mpat(filename, normalizing_gene = NULL, title = "mPAT results",
  pipeline_dir = NULL, max_tail = 300, grouping = NULL,
  plot_count = NULL, plot_tail = NULL)
}
\arguments{
\item{filename}{A counts.csv file as produced by "tail-tools analyse-tail-lengths:".}

\item{normalizing_gene}{ID of housekeeping gene to normalize to. A vector of IDs may be given. NULL means perform no normalization.}

\item{title}{Title of Shiny report.}

\item{pipeline_dir}{Pipeline output directory, if you want to view detailed poly(A) tail length distributions extracted from BAM files.}

\item{max_tail}{Default limit on tail length for tail length distribution plot.}
}
\value{
A shiny appobj. Printing it runs the app.
}
\description{
Shiny report for mPAT results
}

