% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_patseq_heatmap.R
\name{shiny_patseq_heatmap}
\alias{shiny_patseq_heatmap}
\title{Produces detailed heatmap}
\usage{
shiny_patseq_heatmap(datfr, sample_labels = NULL, sample_labels2 = NULL,
  feature_labels = NULL, prefix = "", species = NULL)
}
\arguments{
\item{datfr}{List of dataframes 
Takes a read.grouped.table() as input or a list of four dataframes (more data frames are ok but it only uses these):
Counts - Genewise counts of expression.
Tail - Mean tail length 
Tail_counts - Number of poly-A tails counted.
Annotation - Information regarding the annotation information
     Gene name, chromosome, gene product, biotype etc...}

\item{sample_labels}{Sample labels}

\item{sample_labels2}{Sample labels (second plot)}

\item{feature_labels}{Feature labels}

\item{prefix}{Prefix for plot}

\item{species}{Species of the data. Currently supports Human (Hs), Saccharomyces cerevisiae (Sc), Caenorhabditis elegans (Ce), Mus musculus (Mm)
Now optional. Not entering this disables GO term analysis}
}
\value{
Returns a composable shiny app object
}
\details{
Workhorse function for this package.
}

