% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_repat.R
\name{shiny_repat}
\alias{shiny_repat}
\title{Shiny report for REPAT results}
\usage{
shiny_repat(filename, normalizing_gene = NULL, title = "3'TAP results")
}
\description{
Shiny report for REPAT results
}

