% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_table.R
\name{shiny_table}
\alias{shiny_table}
\title{Shiny dataTable with download button}
\usage{
shiny_table(df = NULL, options = NULL, title = "Table",
  filename = "table.csv", prefix = "")
}
\description{
Shiny dataTable with download button
}

