% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_tail_distribution.R
\name{shiny_tail_distribution}
\alias{shiny_tail_distribution}
\title{Display information about a genomic region}
\usage{
shiny_tail_distribution(peak_names = NULL, sample_normalizer = NULL,
  pipeline_dir, peaks = NULL, max_tail = 300, grouping = NULL,
  prefix = "")
}
\description{
Display information about a genomic region
}

