% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_diff_test.R
\name{shiny_tests}
\alias{shiny_tests}
\title{Present a collection of differential tests}
\usage{
shiny_tests(
  tests,
  cache_prefix = "cache_",
  title = "Differential tests",
  prefix = ""
)
}
\arguments{
\item{tests}{A named list of tests. Each item is a list("function_name", ...arguments...).}
}
\description{
Present a collection of differential tests
}
