% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tail_counts_get_matrix}
\alias{tail_counts_get_matrix}
\title{Extract a column from tail_counts$obs as a matrix}
\usage{
tail_counts_get_matrix(tc, column_name)
}
\description{
Extract a column from tail_counts$obs as a matrix
}
