% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tail_counts_vst}
\alias{tail_counts_vst}
\title{Augment a tail_counts by performing varistran's vst on the counts}
\usage{
tail_counts_vst(tc)
}
\description{
Augment a tail_counts by performing varistran's vst on the counts
}
