% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailtools.R
\docType{package}
\name{tailtools-package}
\alias{tailtools}
\alias{tailtools-package}
\title{Tail Tools R library}
\description{
This is the R library for Tail Tools, and is mostly to do with presenting the output of the Tail Tools python-based pipeline.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Victorian-Bioinformatics-Consortium/tail-tools}
}

}
