% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_test.R
\name{test_diff_exp}
\alias{test_diff_exp}
\title{Test for differential expression}
\usage{
test_diff_exp(counts_filename, min_reads = 10, samples = NULL, design, ...)
}
\arguments{
\item{min_reads}{There must be this many reads for an item to be included in the test, summing over all relevant samples.}
}
\description{
Test for differential expression
}

