% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_test.R
\name{test_diff_tail}
\alias{test_diff_tail}
\title{Test for differential tail length}
\usage{
test_diff_tail(pipeline_dir, design, contrast = NULL, coef1 = NULL,
  coef2 = NULL, min_reads = 10, samples = NULL, title = NULL,
  step = 0.001, fdr = 0.05)
}
\description{
Test for differential tail length
}
