% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_test.R
\name{test_end_shift}
\alias{test_end_shift}
\title{Test for 3' end shift}
\usage{
test_end_shift(
  pipeline_dir,
  design,
  coef1,
  coef2,
  min_reads = 10,
  samples = NULL,
  fdr = 0.05,
  step = 0.001,
  antisense = F,
  colliders = F,
  non_utr = F,
  collapse_utr = F,
  title = NULL
)
}
\arguments{
\item{pipeline_dir}{Pipeline directory.}

\item{design}{Design matrix. Should have a column for expression level in condition 1, and a column for expression level in condition 2.}

\item{coef1}{Column number in the design matrix for expression level in condition 1.}

\item{coef2}{Column number in the design matrix for expression level in condition 2.}

\item{min_reads}{Minimum total reads in order to use a peak.}

\item{samples}{Sample names to be used. Rows of the design matrix should correspond to these samples. Defaults to all samples in the order they were given when the pipeline was run.}

\item{fdr}{False Discovery Rate to maintain.}

\item{step}{Accuracy of "confect" value. Should not need to be changed.}

\item{antisense}{Include most antisense peaks.}

\item{colliders}{In combination with antisense=TRUE, include antisense peaks, even those that are more likely sense peaks for another gene.}

\item{non_utr}{Include non-3'UTR peaks.}

\item{collapse_utr}{Use in combination with non_utr=TRUE. Collapse 3'UTR peaks into a "single peak", to look for shifts to/from non-3'UTR peaks.}

\item{title}{A title for this test.}
}
\description{
Test for 3' end shift
}
