% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_shift.R
\name{end_shift_pipeline}
\alias{end_shift_pipeline}
\title{End shift statistics from tail-tools pipeline output.}
\usage{
end_shift_pipeline(path, condition, group = NULL, ci = 0.95, fdr = T,
  antisense = T, non_utr = T)
}
\arguments{
\item{fdr}{Produce permutation based q values, can be very slow.}
}
\description{
End shift statistics from tail-tools pipeline output.
}

