% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pipeline_bams}
\alias{pipeline_bams}
\title{Get BAM filenames from a Tail Tools pipeline output directory}
\usage{
pipeline_bams(pipeline_dir)
}
\description{
Get BAM filenames from a Tail Tools pipeline output directory
}

