% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_end_shift.R
\name{shiny_end_shift_pipeline}
\alias{shiny_end_shift_pipeline}
\title{Perform an end shift analysis and report results.}
\usage{
shiny_end_shift_pipeline(path)
}
\arguments{
\item{path}{Directory containing tail-tools pipeline output}
}
\description{
Perform an end shift analysis and report results.
}

