% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_test.R
\name{test_diff_exp}
\alias{test_diff_exp}
\title{Test for differential expression}
\usage{
test_diff_exp(
  pipeline_dir,
  design,
  contrast = NULL,
  coef1 = NULL,
  coef2 = NULL,
  min_reads = 10,
  samples = NULL,
  title = NULL,
  step = 0.001,
  fdr = 0.05,
  what = "genewise"
)
}
\arguments{
\item{min_reads}{There must be this many reads for an item to be included in the test, summing over all relevant samples.}
}
\description{
Test for differential expression
}
