% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weitrix_test.R
\name{test_diff_exp}
\alias{test_diff_exp}
\title{Legacy test for differential expression}
\usage{
test_diff_exp(
  pipeline_dir,
  design,
  contrast = NULL,
  coef1 = NULL,
  coef2 = NULL,
  min_reads = 10,
  samples = NULL,
  title = NULL,
  step = 0.001,
  fdr = 0.05,
  what = "genewise"
)
}
\description{
Legacy test for differential expression
}
